requirements.txt
fastapi>=0.111
uvicorn>=0.29
langgraph>=0.1
langchain-google-genai>=1.0
google-generativeai>=0.5
httpx>=0.27
python-multipart>=0.0.9
pydantic>=2.0
pinecone-client>=3.0
python-dotenv>=1.0
sarvamai
